src/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
Django==3.2.11
django-ckeditor==6.4.1
django-cors-headers==3.12.0
django-filter==21.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
drf-writable-nested==0.6.3
email-validator==1.2.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysqlclient==2.1.0
phonenumbers==8.12.49
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
razorpay==1.3.0
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
twilio==7.9.1
urllib3==1.26.9
whitenoise==6.1.0
